% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bme_cv.R
\name{bme_cv}
\alias{bme_cv}
\title{bme_cv function}
\usage{
bme_cv(ch, cs, zh, a, b, model, nugget, sill, range, nsmax, nhmax, type)
}
\arguments{
\item{ch}{matrix of hard data locations}

\item{cs}{matrix of soft data locations}

\item{zh}{vector of hard data}

\item{a}{vector of lower bounds of soft data}

\item{b}{vector of lower bounds of soft data}

\item{model}{string name of covariance or variogram model}

\item{nugget}{a non-negative value}

\item{sill}{a non-negative value}

\item{range}{a non-negative value}

\item{nsmax}{number of soft data locations closer to the estimation location}

\item{nhmax}{number of hard data locations closer to the estimation location}

\item{type}{string name for the type of prediction preferred. Type of
prediction can either be "mean" (posterior mean) or "mode" (posterior
mode)}
}
\value{
Data frame of estimation locations with their corresponding
predictions, residuals and folds.
}
\description{
bme_cv function
}
\details{
Leave-one-out cross validation (LOOCV) function to compute
predictions for ONLY hard data locations.
}
\examples{
data("utah")
ch <- data.matrix(utah[2:10, c("x", "y")])
cs <- data.matrix(utah[68:232, c("x", "y")])
zh <- c(utah[2:10, c("center")])
a <- c(utah[68:232, c("lower")])
b <- c(utah[68:232, c("upper")])
model <- "sph"
nugget <- 0.1184
sill <- 0.3474
range <- 119197
nsmax <- 5
nhmax <- 5
type <- "mean"
bme_cv(ch, cs, zh, a, b, model, nugget, sill, range, nsmax, nhmax, type)

}
