% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{casnowload}
\alias{casnowload}
\title{California Snow Load Data}
\format{
A data frame with 346 rows and 8 columns.

\describe{
\item{STATION}{Name of the snow measuring station}
\item{LATITUDE}{Latitude coordinate position}
\item{LONGITUDE}{Longitude coordinate position}
\item{ELEVATION}{Elevation of the measring station (measured in meters)}
\item{RTSL}{The hard data RTSL value}
\item{LOWER}{The lower endpoint RTSL}
\item{UPPER}{The upper endpoint RTSL}
\item{TYPE}{Type of snow measurement, WESD is direct and SNWD is indirect
measurement. Direct measurements are hard data and have the
lower, upper and center values are the same. Indirect
measurements have LOWER < RTSL < UPPER.}
}
}
\source{
\url{https://www1.ncdc.noaa.gov/pub/data/ghcn/daily/}
}
\usage{
casnowload
}
\description{
A subset of data from the 7964 measurement locations included in the 2020
National Snow Load Study. This data is basically on reliability-targeted
snow loads (RTSL) in the state of California.
}
\keyword{datasets}
