% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extended_range.R
\name{extended_range}
\alias{extended_range}
\title{Computes an extended numeric range that includes all elements from three
numeric vectors: x, y, z. The range is extended by 10\\% of the original
range on both sides}
\usage{
extended_range(zh, a, b)
}
\description{
Computes an extended numeric range that includes all elements from three
numeric vectors: x, y, z. The range is extended by 10\\% of the original
range on both sides
}
\keyword{internal}
