% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.BMEmapping.R
\name{plot.BMEmapping}
\alias{plot.BMEmapping}
\title{Plot Method for BMEmapping Objects}
\usage{
\method{plot}{BMEmapping}(x, ...)
}
\arguments{
\item{x}{An object of class \code{"BMEmapping"}.}

\item{...}{Additional arguments (currently unused).}
}
\value{
A \pkg{ggplot2} object or an arranged collection of plots,
returned invisibly.
}
\description{
Generates graphical displays for objects of class \code{"BMEmapping"}.
The output depends on the object type:
\itemize{
\item Objects returned by \code{bme_map()} produce a spatial plot showing
hard and soft data locations.
\item Objects returned by \code{prob_zk()} produce posterior density plots.
\item Objects returned by \code{bme_predict()} or \code{bme_interval()}
produce spatial prediction maps.
\item Objects returned by \code{bme_cv()} produce diagnostic plots for
evaluating prediction performance.
}
}
