% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{utah}
\alias{utah}
\title{A hard and soft-interval design ground snow load dataset for Utah.}
\format{
A data frame with 232 rows and 5 variables:

 \describe{
 \item{lat}{Latitude coordinate position}
 \item{lon}{Longitude coordinate position}
 \item{center}{The hard data value}
 \item{lower}{The lower endpoint of the soft-interval}
 \item{upper}{The upper endpoint of the soft-interval}
 }
}
\source{
* [The 2020 National Snow Load Study](https://doi.org/10.26077/200k-pr86)
}
\usage{
utah
}
\description{
A dataset containing the 67 hard data and 165 soft-interval data used in the
analysis of Duah et. al. (2025). The 232 measurement locations
included in the dataset are taken from The 2020 National Snow Load Study
(Bean et. al., 2021).
}
\keyword{datasets}
