% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{utahsnowload}
\alias{utahsnowload}
\alias{utsnowload}
\title{A detrended reliability-targeted design ground snow loads in Utah}
\format{
A data frame with 232 rows and 5 variables:

\describe{
\item{latitude}{Latitude coordinate position}
\item{longitude}{Longitude coordinate position}
\item{hard}{The hard data value}
\item{lower}{The lower endpoint of the soft-interval}
\item{upper}{The upper endpoint of the soft-interval}
}
}
\source{
\doi{10.1016/j.spasta.2025.100894}
}
\usage{
utsnowload
}
\description{
This dataset contains detrended reliability-targeted design ground snow load
measurements from 232 locations in state of Utah. Of these, 65 sites report
precise measurements, treated as hard data, while the remaining 167 sites
report imprecise measurements, represented as interval (soft) data. The
dataset is structured such that the first 67 rows contain hard (point)
measurements, and the remaining rows represent soft data using lower and
upper interval bounds. For a detailed explanation of the dataset and its use,
refer to the related version described in Duah et al. (2025)
\url{doi:10.1016/j.spasta.2025.100894}
}
\keyword{datasets}
