% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bme_predict.R
\name{bme_predict}
\alias{bme_predict}
\title{bme_predict function}
\usage{
bme_predict(
  x,
  ch,
  cs,
  zh,
  a,
  b,
  model,
  nugget,
  sill,
  range,
  nsmax,
  nhmax,
  type
)
}
\arguments{
\item{x}{matrix of estimation locations}

\item{ch}{matrix of hard data locations}

\item{cs}{matrix of soft data locations}

\item{zh}{vector of hard data}

\item{a}{vector of lower bounds of soft data}

\item{b}{vector of lower bounds of soft data}

\item{model}{string name of covariance or variogram model}

\item{nugget}{a non-negative value}

\item{sill}{a non-negative value}

\item{range}{a non-negative value}

\item{nsmax}{number of soft data locations closer to the estimation location}

\item{nhmax}{number of hard data locations closer to the estimation location}

\item{type}{string name for the type of prediction preferred. Type of
prediction can either be "mean" (posterior mean) or "mode" (posterior
mode)}
}
\value{
Data frame of estimation locations with their corresponding BME
predictions
}
\description{
bme_predict function
}
\details{
Compute the BME predictions for estimation locations.
}
\examples{
data("utah")
x <- data.matrix(utah[1, c("x", "y")])
ch <- data.matrix(utah[2:67, c("x", "y")])
cs <- data.matrix(utah[68:232, c("x", "y")])
zh <- c(utah[2:67, c("center")])
a <- c(utah[68:232, c("lower")])
b <- c(utah[68:232, c("upper")])
model <- "sph"
nugget <- 0.1184
sill <- 0.3474
range <- 119197
nsmax <- 5
nhmax <- 10
type <- "mean"
bme_predict(x, ch, cs, zh, a, b, model, nugget, sill, range, nsmax, nhmax,
            type)

}
