% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.BMEmapping.R
\name{plot.BMEmapping}
\alias{plot.BMEmapping}
\title{Plot Method for BMEmapping Objects (ggplot2)}
\usage{
\method{plot}{BMEmapping}(x, ...)
}
\arguments{
\item{x}{An object of class \code{"BMEmapping"}.}

\item{...}{Additional arguments (currently unused).}
}
\description{
Uses ggplot2 to generate diagnostic or posterior density plots for objects
of class \code{"BMEmapping"} returned by \code{bme_cv()} or \code{prob_zk()}.
}
