% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prob_zk.R
\name{prob_zk}
\alias{prob_zk}
\title{Posterior density function}
\usage{
prob_zk(
  x,
  ch,
  cs,
  zh,
  a,
  b,
  model,
  nugget,
  sill,
  range,
  nsmax = 10,
  nhmax = 10
)
}
\arguments{
\item{x}{matrix of estimation locations. Cannot exceed 10 locations.}

\item{ch}{matrix of hard data locations}

\item{cs}{matrix of soft data locations}

\item{zh}{vector of hard data}

\item{a}{vector of lower bounds of soft data}

\item{b}{vector of lower bounds of soft data}

\item{model}{string name of covariance or variogram model}

\item{nugget}{a non-negative value}

\item{sill}{a non-negative value}

\item{range}{a non-negative value}

\item{nsmax}{number of soft data locations closer to the estimation location}

\item{nhmax}{number of hard data locations closer to the estimation location}
}
\value{
A two column matrix of zk and associated probabilities
}
\description{
Posterior density function
}
\details{
Compute the posterior densities of assumed zki values
}
\examples{
data("utah")
x <- data.matrix(utah[1, c("lat", "lon")])
ch <- data.matrix(utah[2:67, c("lat", "lon")])
cs <- data.matrix(utah[68:232, c("lat", "lon")])
zh <- c(utah[2:67, c("center")])
a <- c(utah[68:232, c("lower")])
b <- c(utah[68:232, c("upper")])
model <- "sph"
nugget <- 0.1184
sill <- 0.3474
range <- 119197
nsmax <- 5
nhmax <- 10
prob_zk(x, ch, cs, zh, a, b, model, nugget, sill, range, nsmax, nhmax)

}
