% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{utah}
\alias{utah}
\title{A projected hard and soft-interval design ground snow load dataset for Utah.}
\format{
A data frame with 232 rows and 7 variables:

\describe{
\item{latitude}{Latitude coordinate position}
\item{longitude}{Longitude coordinate position}
\item{x}{Projected latitude coordinate}
\item{y}{Projected longitude coordinate}
\item{center}{The hard data value}
\item{lower}{The lower endpoint of the soft-interval}
\item{upper}{The upper endpoint of the soft-interval}
}
}
\source{
\url{https://www1.ncdc.noaa.gov/pub/data/ghcn/daily/}
}
\usage{
utah
}
\description{
A dataset comprising 65 hard data points and 167 soft-interval data points,
used in the analysis by Duah et al. (2025). The dataset includes 248
measurement locations, which are derived from the 2020 National Snow Load
Study (Bean et al., 2021). For a comprehensive understanding of the dataset
and its usage, please refer to Duah et al. (2025).
}
\keyword{datasets}
